+++ packages/xlucene-evaluator/src/document-matcher/type-manager/index.ts
