+++ packages/xlucene-evaluator/src/document-matcher/type-manager/types/base.ts
