+++ packages/xlucene-evaluator/src/document-matcher/type-manager/types/string.ts
