+++ packages/xlucene-evaluator/src/interfaces.ts
