+++ packages/xlucene-evaluator/src/peg.ts
@@ -5,10 +5,10 @@ let parser: PegEngine;
 
 try {
      // @ts-ignore because the types don't exist
-    parser = require('../../peg/peg_engine');
+    parser = require('../peg/peg_engine');
 } catch (err) {
     // @ts-ignore because the types don't exist
-    parser = require('../../../peg/peg_engine');
+    parser = require('../../peg/peg_engine');
 }
 
 export default parser;
